+++ src/savefile.h
@@ -100,6 +100,7 @@ int rd_player(void);
 int rd_ignore(void);
 int rd_misc(void);
 int rd_player_hp(void);
+int rd_player_spells_1(void);
 int rd_player_spells(void);
 int rd_gear(void);
 int rd_stores(void);